Rethought action listing.
Previously, the intention was to get a list of all possible actions,
then execute them all in turn.  In reality, that's not at all how a game
goes.  One determines all possible actions, and does one of them, then
repeats the process.

Now, `GetActions()` returns a slice of actions, expecting one and only
one to be executed.